Added SDK >= 3.0.0 requirement.
Fixed getContact and getRawContactByRawId behaviour for when there is no contact to return.
The signature would lead to assume that if there is no contact to return the awaited value would simply be null, but the prior behavior would always return Future<Contact> or throw an exception.

Fixed behavior to match signature.